refactor(v1): the run types live with the episode they describe
They were written when the run was a field on Trace. #2244 moved the
field to Episode; the types stayed behind, and Trace has not referred to
one since.
