+++ verifiers/v1/episode.py
@@ -1,14 +1,14 @@
 """The episode — one run's traces plus their shared standing, whole."""
 
 import uuid
-from typing import Any, Generic
+from typing import Annotated, Any, Generic, Literal
 
 from pydantic import BaseModel, Field
 
 from verifiers.v1.configs.agent import WireAgentConfig
 from verifiers.v1.state import State, StateT
 from verifiers.v1.task import DataT, WireTaskData
-from verifiers.v1.trace import AgentConfigT, Error, RunInfo, Trace
+from verifiers.v1.trace import AgentConfigT, Error, Trace
 from verifiers.v1.types import Usage
 
 
@@ -27,6 +27,91 @@ class GroupInfo(BaseModel):
 
     size: int = 1
     """How many episodes were planned in it."""
+
+
+class EvalRunInfo(BaseModel):
+    """A standalone eval: a model measured once, against nothing that is training."""
+
+    type: Literal["eval"] = "eval"
+
+    id: str
+
+
+class PolicySpan(BaseModel):
+    """The live policy versions an episode was generated across — `TimeSpan`'s shape, measured in
+    policy updates."""
+
+    start: int = 0
+    """The live version when generation began."""
+
+    end: int = 0
+    """The live version when it finished."""
+
+    @property
+    def drift(self) -> int:
+        return max(0, self.end - self.start)
+
+
+class TrainMetadata(BaseModel):
+    """An episode a training run trains on."""
+
+    type: Literal["train"] = "train"
+
+    step: int | None = None
+    """The batch window collecting when it landed, which is not known until it does."""
+
+    def off_policy_steps(self, policy: PolicySpan) -> int | None:
+        """How far behind the policy being trained the generating policy was — queue time
+        included, since `step` is the window it landed in, not the one it left."""
+        if self.step is None:
+            return None
+        return max(0, (self.step - 1) - policy.start)
+
+
+class EvalMetadata(BaseModel):
+    """An episode a training run measures itself with."""
+
+    type: Literal["eval"] = "eval"
+
+    step: int
+    """The step whose eval produced it, known from the moment it is dispatched."""
+
+    def off_policy_steps(self, policy: PolicySpan) -> int:
+        """How far off the policy it measured drifted while it ran. Nothing trains on an eval, so
+        there is no policy for it to be behind; what makes it off-policy is the policy moving under
+        it, leaving it a measurement of a blend rather than of the version it started on.
+
+        Its `step` cannot answer that — it is fixed when the eval is dispatched, so a slow eval that
+        outlives several updates would still look on-policy."""
+        return policy.drift
+
+
+EpisodeMetadata = Annotated[TrainMetadata | EvalMetadata, Field(discriminator="type")]
+"""What an episode is to the training run it belongs to."""
+
+
+class TrainRunInfo(BaseModel):
+    """A training run. Its episodes are the ones it trains on and the ones it evaluates along the
+    way: one run, one id, and `metadata` says which of the two an episode is."""
+
+    type: Literal["train"] = "train"
+
+    id: str
+
+    metadata: EpisodeMetadata = Field(default_factory=TrainMetadata)
+
+    policy: PolicySpan | None = None
+    """`None` when it was not generated from the live policy at all, as with a frozen sampler."""
+
+    @property
+    def off_policy_steps(self) -> int | None:
+        if self.policy is None:
+            return None
+        return self.metadata.off_policy_steps(self.policy)
+
+
+RunInfo = Annotated[EvalRunInfo | TrainRunInfo, Field(discriminator="type")]
+"""The run an episode belongs to, discriminated on `type`."""
 
 
 class Episode(BaseModel, Generic[DataT, StateT, AgentConfigT]):
